Fix sorting by column.
When overriding DrawRows, we forgot to check for pawnListDirty. This meant the pawn list was never rebuild while the tab was open.

Since pawnListDirty is private, and Notify_PawnsChanged is not virtual, we had to use reflection to check for the dirty flag.

In addition, BuildPawnList forgot to turn the dirty flag off. This has been fixed, so the list won't continuously update anymore.