Add option to delete favourite.
There was never an option to delete favourites, and since they are kept in the save, the favourites list could become cluttered.

This commit adds a delete option that deletes the favourite from the list, as well as de-assigns it from pawns. It also tweaks the list of assignable favourites to not include the currently assigned favourite.

In addition it fully comments out the no longer used NaturalPriorities Dialog and accompanying MapCompontent.