refactor storage of priorities
Completely reworked how priorities are stored. Instead of using defmaps where only the priorities are stored and the workgiverdefs are assumed to stay constant, we now store a list of workgiverdefs with the priorities to ascertain ordering, and match priorities to current workgiverdefs upon loading.

In addition, we no longer use a list of lists for storing priorities, but a much more compressed block of digits per pawn, where each line is an hour and each column a workgiver.

Migrating from the old to new format is done by setting a 'newFormat' flag in the save game, and should work out-of-the-box.