Changing an entire Work Type will not overwrite the Work Giver priorities. i.e. incrementing a column or job increments all sub-jobs, instead of writing the new value from the work type.
There are potential issues and questions of design/intent. Incrementing/Decrementing a column is now bounded by the min and max sub-priority, so if one giver is at one, other won't be decremented. This is done to preserve the ranking, although it then means some won't decrement to 1. Incrementing a single pawn's worktype now increments all workgivers, but with no wrap-around checks, so some high priority tasks are blanked while the other workgivers are still active. This is all okay for me, (so many priority levels) but there could be a config option to overwrite giver priorities so that other givers may rise.

Now, there needed to be new Pawn_Extensions since wrap-around checks occur differently: In addition to GetPriority(worktype), there's GetMinPriority, GetMaxPriority, AnyGiverMissingPriority. Some occurences of GetPriority(worktype) were changed if it looked right, some were not. These should be reviewed.